Add delay to LED changes
Previously, an apparent hardware delay was causing undefined behavior
when the button was pressed (light turning off, flashing the wrong
color, etc.) This code adds a sleep to all color change requests, which
gives the hardware time to respond and prevents the undefined behavior.
There doesn't seem to be a way to get the status from the light, so it
ended up being a somewhat arbitrary value choice.